fix: 🐛 moved some code around
