+++ src/components/availability/group-availability.tsx
@@ -9,84 +9,6 @@ import { cn } from "@/lib/utils";
 import type { ZotDate } from "@/lib/zotdate";
 import { useAvailabilityStore } from "@/store/useAvailabilityStore";
 
-// to fit api format, am/pm is needed
-function formatRange(start: Date, end: Date) {
-	const format = (d: Date) => {
-		let hours = d.getHours();
-		const minutes = d.getMinutes().toString().padStart(2, "0");
-
-		const ampm = hours >= 12 ? "pm" : "am";
-		hours = hours % 12;
-		if (hours === 0) hours = 12;
-
-		return `${hours}:${minutes}${ampm}`;
-	};
-	const rangeEnd = new Date(end.getTime() + 15 * 60 * 1000);
-	return `${format(start)}-${format(rangeEnd)}`;
-}
-
-export function getBestTimeRanges(availabilityDates: any[]) {
-	let max = 0;
-
-	availabilityDates.forEach((date) => {
-		Object.values(date.groupAvailability).forEach((memberIds) => {
-			max = Math.max(max, (memberIds as string[]).length);
-		});
-	});
-
-	const timestamps: string[] = [];
-
-	availabilityDates.forEach((date) => {
-		Object.entries(date.groupAvailability).forEach(
-			([timestamp, memberIds]: [string, unknown]) => {
-				if ((memberIds as string[]).length === max && max > 0) {
-					timestamps.push(timestamp);
-				}
-			},
-		);
-	});
-
-	if (!timestamps.length) return [];
-	const sorted = [...timestamps].sort();
-	const results: { date: string; time: string }[] = [];
-
-	let start = new Date(sorted[0]);
-	let prev = start;
-
-	for (let i = 1; i < sorted.length; i++) {
-		const curr = new Date(sorted[i]);
-
-		const diffMinutes = (curr.getTime() - prev.getTime()) / 60000;
-
-		const sameDay =
-			curr.toISOString().split("T")[0] === prev.toISOString().split("T")[0];
-
-		if (diffMinutes !== 15 || !sameDay) {
-			results.push({
-				date: normalizeDate(start),
-				time: formatRange(start, prev),
-			});
-			start = curr;
-		}
-		prev = curr;
-	}
-
-	results.push({
-		date: normalizeDate(start),
-		time: formatRange(start, prev),
-	});
-
-	return results;
-}
-
-function normalizeDate(date: Date) {
-	const year = date.getFullYear();
-	const month = String(date.getMonth() + 1).padStart(2, "0");
-	const day = String(date.getDate()).padStart(2, "0");
-
-	return `${year}-${month}-${day}`;
-}
-
 export const getTimestampFromBlockIndex = (
 	blockIndex: number,
 	zotDateIndex: number,
