+++ src/lib/meetings/utils.ts
@@ -5,7 +5,7 @@ interface TimeInterval {
 	to: string;
 }
 
-function formatLocalDateKey(date: Date): string {
+export function formatLocalDateKey(date: Date): string {
 	const yyyy = date.getFullYear();
 	const mm = String(date.getMonth() + 1).padStart(2, "0");
 	const dd = String(date.getDate()).padStart(2, "0");
