+++ src/lib/rooms/get-rooms.ts
@@ -27,6 +27,10 @@ export async function fetchStudyRooms({
 	};
 	const params = new URLSearchParams(query).toString();
 	const apiUrl = `https://anteaterapi.com/v2/rest/studyRooms?${params}`;
+
+	// temporary logging for debugging, reproducing api results
+	console.log(`Fetching study rooms with URL: ${apiUrl}`);
+
 	const res = await fetch(apiUrl);
 	if (!res.ok) {
 		throw new Error(`API error: ${res.status} ${res.statusText}`);
